Added new SAMFileReader.query method that enables querying of multiple intervals in a single invocation.  In many cases this will be much more efficient than querying each interval separately, because due to the limitations of BAM index a query can result in many records read from the file that do not satisfy the query parameters.  In multiple query invocations, the same regions of the file may be read more than once, whereas with the new method each candidate region of the file will be read only once.
SamLocusIterator now uses this method.  It used to be the case that for some inputs, having SamLocusIterator use an index was less efficient than a linear scan through the whole file.  It should now be the case that using the index will always be at least as efficient as a linear scan, an generally using the index will be much more efficient.

git-svn-id: http://svn.code.sf.net/p/picard/code/trunk@1650 8bce386e-0d6e-43e0-b1ca-608082c0432d